Crate stylo_static_prefs

Macros

pref

Returns the value of a preference exposed to the style crate. If the embedder has not set a value for it, this returns the default value of the preference

set_pref

Statics

BOOLS

INTEGERS