stylo_static_prefs

Macro pref

macro_rules! pref {
    ("dom.select.customizable_select.enabled") => { ... };
    ("dom.viewTransitions.cross-document.enabled") => { ... };
    ("gfx.font_rendering.opentype_svg.enabled") => { ... };
    ("layout.columns.enabled") => { ... };
    ("layout.container-queries.enabled") => { ... };
    ("layout.css.alpha-color-function.enabled") => { ... };
    ("layout.css.anchor-positioning.enabled") => { ... };
    ("layout.css.appearance-base.enabled") => { ... };
    ("layout.css.at-scope.enabled") => { ... };
    ("layout.css.attr.enabled") => { ... };
    ("layout.css.background-clip.border-area.enabled") => { ... };
    ("layout.css.basic-shape-shape.enabled") => { ... };
    ("layout.css.color-mix-multi-color.enabled") => { ... };
    ("layout.css.content.alt-text.enabled") => { ... };
    ("layout.css.contrast-color.enabled") => { ... };
    ("layout.css.cross-fade.enabled") => { ... };
    ("layout.css.custom-media.enabled") => { ... };
    ("layout.css.ellipse-corners.enabled") => { ... };
    ("layout.css.fit-content-function.enabled") => { ... };
    ("layout.css.font-palette.enabled") => { ... };
    ("layout.css.font-tech.enabled") => { ... };
    ("layout.css.gradient-color-interpolation-method.enabled") => { ... };
    ("layout.css.grid-template-masonry-value.enabled") => { ... };
    ("layout.css.has-selector.enabled") => { ... };
    ("layout.css.light-dark.images.enabled") => { ... };
    ("layout.css.line-clamp.enabled") => { ... };
    ("layout.css.margin-rules.enabled") => { ... };
    ("layout.css.marker.restricted") => { ... };
    ("layout.css.motion-path-url.enabled") => { ... };
    ("layout.css.nth-child-of.enabled") => { ... };
    ("layout.css.overflow-moz-hidden-unscrollable.enabled") => { ... };
    ("layout.css.progress-function.enabled") => { ... };
    ("layout.css.properties-and-values.enabled") => { ... };
    ("layout.css.relative-color-syntax.enabled") => { ... };
    ("layout.css.revert-rule.enabled") => { ... };
    ("layout.css.scroll-driven-animations.enabled") => { ... };
    ("layout.css.scroll-state.enabled") => { ... };
    ("layout.css.starting-style-at-rules.enabled") => { ... };
    ("layout.css.stretch-size-keyword.enabled") => { ... };
    ("layout.css.style-queries.enabled") => { ... };
    ("layout.css.system-ui.enabled") => { ... };
    ("layout.css.text-decoration-inset-percentage.enabled") => { ... };
    ("layout.css.tree-counting-functions.enabled") => { ... };
    ("layout.css.webkit-fill-available.all-size-properties.enabled") => { ... };
    ("layout.css.webkit-fill-available.enabled") => { ... };
    ("layout.grid.enabled") => { ... };
    ("layout.unimplemented") => { ... };
    ("layout.variable_fonts.enabled") => { ... };
    ("layout.writing-mode.enabled") => { ... };
    ("layout.css.outline-offset.snapping") => { ... };
    ("layout.css.stylo-local-work-queue.in-main-thread") => { ... };
    ("layout.css.stylo-local-work-queue.in-worker") => { ... };
    ("layout.css.stylo-work-unit-size") => { ... };
    ("layout.threads") => { ... };
}

Returns the value of a preference exposed to the style crate. If the embedder has not set a value for it, this returns the default value of the preference